Corrected errors in these files

+++ core-java/src/main/java/org/jn/core/java/student/Anas/Sirelkhatim/Arithmetic_Operation/Problem26.java
@@ -6,7 +6,7 @@ public class Problem26 {
         
         short s = 10;
         int a = s;
-        s = s + 10;
+        s = (short) (s + 10);
         System.out.println(s + "," + a);
 
     }
